Rename question-service model filename
